Still developing

+++ src/main/java/com/Toiema/StudentRegistration/repository/studentRepository.java
@@ -4,7 +4,7 @@ import com.Toiema.StudentRegistration.entity.Student;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface studentRepository extends JpaRepository<Student,Long> {
+public interface studentRepository extends JpaRepository<Student,Integer> {
 
     // no coding here
 }
